refactor: change backend structure
