+++ backend/cmd/generate.go
@@ -1,0 +1,1 @@
+package cmd
