+++ backend/go.mod
@@ -1,3 +1,9 @@
-module github/thomasoca/cv-generator/backend
+module github.com/thomasoca/cv-generator/backend
 
-go 1.16
+go 1.19
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/cobra v1.6.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
