+++ backend/pkg/image/image_test.go
@@ -1,4 +1,4 @@
-package main
+package image
 
 import (
 	"log"
