+++ backend/pkg/utils/utils.go
@@ -1,0 +1,16 @@
+package utils
+
+import (
+	"io/ioutil"
+	"log"
+	"os"
+)
+
+func JsonInput(fname string) []byte {
+	jsonFile, err := os.Open(fname)
+	if err != nil {
+		log.Println(err)
+	}
+	byteValue, _ := ioutil.ReadAll(jsonFile)
+	return byteValue
+}
