+++ backend/scripts/setup_latex.sh
